+++ projects/schematics/src/ng-add/versions.ts
@@ -1,4 +1,4 @@
-export const angularFontawesomeVersion = '^1.0.0';
+export const angularFontawesomeVersion = '^2.0.0';
 export const iconPackVersion = '^6.7.1';
 
 export const v5 = {