+++ src/lib/icon/duotone-icon.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { IconDefinition as CoreIconDefinition, IconParams } from '@fortawesome/fontawesome-svg-core';
 import { IconDefinition, IconProp } from '../types';
 import { FaIconComponent } from './icon.component';
@@ -6,6 +6,7 @@ import { FaIconComponent } from './icon.component';
 @Component({
   selector: 'fa-duotone-icon',
   template: ``,
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class FaDuotoneIconComponent extends FaIconComponent {
   /**
@@ -15,7 +16,7 @@ export class FaDuotoneIconComponent extends FaIconComponent {
    *
    * @default false
    */
-  @Input() swapOpacity?: 'true' | 'false' | boolean;
+  readonly swapOpacity = input<'true' | 'false' | boolean>();
 
   /**
    * Customize the opacity of the primary icon layer.
@@ -23,7 +24,7 @@ export class FaDuotoneIconComponent extends FaIconComponent {
    *
    * @default 1.0
    */
-  @Input() primaryOpacity?: string | number;
+  readonly primaryOpacity = input<string | number>();
 
   /**
    * Customize the opacity of the secondary icon layer.
@@ -31,7 +32,7 @@ export class FaDuotoneIconComponent extends FaIconComponent {
    *
    * @default 0.4
    */
-  @Input() secondaryOpacity?: string | number;
+  readonly secondaryOpacity = input<string | number>();
 
   /**
    * Customize the color of the primary icon layer.
@@ -39,7 +40,7 @@ export class FaDuotoneIconComponent extends FaIconComponent {
    *
    * @default CSS inherited color
    */
-  @Input() primaryColor?: string;
+  readonly primaryColor = input<string>();
 
   /**
    * Customize the color of the secondary icon layer.
@@ -47,7 +48,7 @@ export class FaDuotoneIconComponent extends FaIconComponent {
    *
    * @default CSS inherited color
    */
-  @Input() secondaryColor?: string;
+  readonly secondaryColor = input<string>();
 
   protected findIconDefinition(i: IconProp | IconDefinition): CoreIconDefinition | null {
     const definition = super.findIconDefinition(i);
@@ -67,7 +68,8 @@ export class FaDuotoneIconComponent extends FaIconComponent {
   protected buildParams(): IconParams {
     const params = super.buildParams();
 
-    if (this.swapOpacity === true || this.swapOpacity === 'true') {
+    const swapOpacity = this.swapOpacity();
+    if (swapOpacity === true || swapOpacity === 'true') {
       if (Array.isArray(params.classes)) {
         params.classes.push('fa-swap-opacity');
       } else if (typeof params.classes === 'string') {
@@ -80,17 +82,17 @@ export class FaDuotoneIconComponent extends FaIconComponent {
     if (params.styles == null) {
       params.styles = {};
     }
-    if (this.primaryOpacity != null) {
-      params.styles['--fa-primary-opacity'] = this.primaryOpacity.toString();
+    if (this.primaryOpacity() != null) {
+      params.styles['--fa-primary-opacity'] = this.primaryOpacity().toString();
     }
-    if (this.secondaryOpacity != null) {
-      params.styles['--fa-secondary-opacity'] = this.secondaryOpacity.toString();
+    if (this.secondaryOpacity() != null) {
+      params.styles['--fa-secondary-opacity'] = this.secondaryOpacity().toString();
     }
-    if (this.primaryColor != null) {
-      params.styles['--fa-primary-color'] = this.primaryColor;
+    if (this.primaryColor() != null) {
+      params.styles['--fa-primary-color'] = this.primaryColor();
     }
-    if (this.secondaryColor != null) {
-      params.styles['--fa-secondary-color'] = this.secondaryColor;
+    if (this.secondaryColor() != null) {
+      params.styles['--fa-secondary-color'] = this.secondaryColor();
     }
 
     return params;