+++ src/lib/stack/stack.component.ts
@@ -1,36 +1,29 @@
-import { Component, ElementRef, Input, OnChanges, OnInit, Renderer2, SimpleChanges } from '@angular/core';
+import { Component, input, computed, ChangeDetectionStrategy } from '@angular/core';
 import { SizeProp } from '@fortawesome/fontawesome-svg-core';
 
 @Component({
   selector: 'fa-stack',
   template: `<ng-content></ng-content>`,
+  host: {
+    '[class]': 'classes()',
+  },
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class FaStackComponent implements OnInit, OnChanges {
+export class FaStackComponent {
   /**
    * Size of the stacked icon.
    * Note that stacked icon is by default 2 times bigger, than non-stacked icon.
    * You'll need to set size using custom CSS to align stacked icon with a
    * simple one. E.g. `fa-stack { font-size: 0.5em; }`.
    */
-  @Input() size?: SizeProp;
+  readonly size = input<SizeProp>();
 
-  constructor(
-    private renderer: Renderer2,
-    private elementRef: ElementRef,
-  ) {}
-
-  ngOnInit() {
-    this.renderer.addClass(this.elementRef.nativeElement, 'fa-stack');
-  }
-
-  ngOnChanges(changes: SimpleChanges) {
-    if ('size' in changes) {
-      if (changes.size.currentValue != null) {
-        this.renderer.addClass(this.elementRef.nativeElement, `fa-${changes.size.currentValue}`);
-      }
-      if (changes.size.previousValue != null) {
-        this.renderer.removeClass(this.elementRef.nativeElement, `fa-${changes.size.previousValue}`);
-      }
-    }
-  }
+  readonly classes = computed(() => {
+    const sizeValue = this.size();
+    const sizeClass = sizeValue ? { [`fa-${sizeValue}`]: true } : {};
+    return {
+      ...sizeClass,
+      'fa-stack': true,
+    };
+  });
 }